pyproject.toml
[project]
name = "doc-lm"
version = "0.1.0"
description = "An LLM-powered document chatbot agent."
readme = "README.md"
authors = [
    { name = "segfaultering", email = "261435586+segfaultering@users.noreply.github.com" }
]
requires-python = "<3.14, >=3.13"
dependencies = [
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.setuptools.packages.find]
where = ["src"]
